+++ packages/article/article-body/article-body-pullquote-wrapper.web.js
@@ -1,0 +1,21 @@
+import React from "react";
+import PropTypes from "prop-types";
+import {
+  PullQuoteContainer,
+  PullQuoteResp
+} from "../styles/article-body/responsive";
+
+const PullQuoteWrapper = ({ children }) => (
+  <PullQuoteContainer>
+    <PullQuoteResp>{children}</PullQuoteResp>
+  </PullQuoteContainer>
+);
+
+PullQuoteWrapper.propTypes = {
+  children: PropTypes.oneOfType([
+    PropTypes.element,
+    PropTypes.arrayOf(PropTypes.element)
+  ]).isRequired
+};
+
+export default PullQuoteWrapper;
